requirements.txt
torch==2.9.1
torchvision==0.17.2
torchaudio==2.9.1

transformers==4.49.0
huggingface-hub==0.36.0

sentencepiece==0.2.0
tokenizers>=0.19.0
accelerate>=0.29.0

pillow==10.3.0
gradio==4.44.1
tqdm==4.66.4
numpy==1.26.4
protobuf==5.27.0
pydantic==2.10.6
# For Florence's remote code loading
safetensors>=0.4.2

# Optional but recommended
colorama==0.4.6
einops==0.8.1
timm==1.0.22

# Additional for full functionality (Q&A, voice, translation, image processing)
gTTS==2.5.3
SpeechRecognition==3.14.4
googletrans==4.0.0-rc1
opencv-python==4.10.0.84
tiktoken==0.7.0
regex==2024.7.24
ftfy==6.2.0